pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "abm_check"
version = "1.0.0"
description = "ABEMA番組情報管理CLIツール"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "yt-dlp>=2024.0.0",
    "PyYAML>=6.0",
    "click>=8.1.0",
    "python-dateutil>=2.8.0",
]

[project.scripts]
abm_check = "abm_check.cli.main:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=abm_check --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
